Growth_Mindset_Challenge/The_Data_Sweeper/requirements.txt
pandas
xlsxwriter
fpdf2
matplotlib